projects/hyperskin/requirements.txt
# --------- python --------- #
numpy
matplotlib
seaborn
pandas
scipy
ipykernel

# --------- pytorch --------- #
# torch
# torchaudio
# torchvision
pytorch-lightning
jsonargparse[signatures]
torchmetrics
segmentation-models-pytorch

# --------- loggers --------- #
# tensorboard
wandb
# neptune-client
# mlflow
# comet-ml

# --------- custom --------- #

# --------- misc --------- #
tqdm
rich
pyrootutils
python-dotenv
debugpy
umap-learn
tabulate

# --------- data --------- #
albumentationsx
timm
scikit-learn
omegaconf
gdown

# --------- image processing --------- #
scikit-image
opencv-python

# interface
gradio
lpips
torchinfo